SQL-1036: Add helper function to verify SQL diagnostic records
