+++ odbc/Cargo.toml
@@ -32,4 +32,4 @@ thiserror = "1"
 
 [lib]
 name = "mongoodbc"
-crate-type = ["cdylib"]
+crate-type = ["cdylib","lib"]
